Data structure API assumption
Without assuming suitable data structure availability large portion of the code is filled with implementation of these data structures and it's hard to review the actual code logic. Most of these data structures are available in PG code so the candidates if hired will not be expected to implement those.