Update dictionary.cpp
The first version of code has been overwritten in order to address the 2 comments.
I understand this code can still be refined further. Also, this is a workable code, which can be compiled and tested with different options.
Usage-
1. to insert - exename insert word
2. to search - exename search word
3. to delete - exename delete word
4. to destroy shared memory - exename exit anyword.
I know that exit command doesn't need any word as such, but this needs to be revised.
Thanks.